+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Changelog
 
-All notable changes to `nova-system-info-card` will be documented in this file
+All notable changes to `nova-system-info-card` will be documented in this file.
+
+## 1.0.2 - 2022-04-07
+
+- Fix database driver check.
 
 ## 1.0.1 - 2019-12-02
 
-- UI changes to fit in a single row
+- UI changes to fit in a single row.
 
 ## 1.0.0 - 2018-08-23
 
-- Initial release
+- Initial release.